AMI Protocol Compliancy Update
The implementation of AMI protocol layer is now fully
compliant. A full description of the protocol as is
implemented and used will be added later.

The last remaining wart had been the handling of the

    Response: Follows
    Priviledge: PRIV
    ActionId: NUMBER
    Some arbitrary message that can span multiple lines
    but eventually ends like follows.
    --END COMMAND--

type of messages. These are now handled properly,
so commands like:

    Action: Command
    Command: core show version

all show their responses properly. Note that these
were problems before this commit, b/cos the previous
implementation (before this commit), assumed every
Response had a format like:

    Response: "Error" | "Success"
    ActionId: NUMBER
    Message: Some detailed Message

This now is fixed.